+++ src/app/[locale]/(logged-in)/data-custodian/profile/[tabId]/components/ProjectUserModalDetails/components/Training.tsx
