+++ src/app/[locale]/(logged-in)/data-custodian/profile/[tabId]/components/ProjectUserModalDetails/index.ts
@@ -1,0 +1,3 @@
+import ProjectUserModalDetails from "./ProjectUserModalDetails";
+
+export default ProjectUserModalDetails;
