+++ src/app/[locale]/(logged-in)/data-custodian/profile/[tabId]/components/ResearcherDetailsModal/ResearcherDetailsModal.tsx
