--- src/app/[locale]/(logged-in)/data-custodian/projects/components/Sections/index.ts
@@ -1,3 +1,0 @@
-import Sections from "./Sections";
-
-export default Sections;
