--- src/app/[locale]/(logged-in)/data-custodian/projects/components/UserModalDetails/index.ts
@@ -1,3 +1,0 @@
-import UserModalDetails from "./UserModalDetails";
-
-export default UserModalDetails;
