+++ src/components/Legend/Legend.tsx
@@ -1,5 +1,5 @@
 import Text from "@/components/Text";
-import { Box } from "@mui/material";
+import { Box, SxProps } from "@mui/material";
 import { ReactNode } from "react";
 
 export interface LegendProps {
@@ -7,11 +7,12 @@ export interface LegendProps {
     text: ReactNode;
     icon: ReactNode;
   }[];
+  boxSx?: SxProps;
 }
 
-export default function Legend({ items }: LegendProps) {
+export default function Legend({ items, boxSx }: LegendProps) {
   return (
-    <Box sx={{ display: "flex", gap: 2, flexWrap: "wrap" }}>
+    <Box sx={{ display: "flex", gap: 2, flexWrap: "wrap", ...boxSx }}>
       {items.map(({ icon, text }) => (
         <Text iconSize="40px" startIcon={icon}>
           {text}
